Move `OutboundRemoveAgentPayload` to `stateful::message`
